Remove the activemodel dependency
We used it for casting the value to a defined type, but since we only
support simple types like integers and strings, we can better use things
like #to_i and #to_s.

The other data type we support are other classes that include
JsonPathAttribute, so we can call .parse on those with the parsed values.